+++ release/b/balochi_inpage/source/help/balochi_inpage.php
@@ -1,15 +1,15 @@
 <?php 
-  $pagename = 'Balochi InPage Keyboard Help';
-  $pagetitle = 'Balochi InPage Keyboard Help';
-  require_once('header.php');
-
-
+$pagename = 'Balochi InPage Keyboard Help';
+$pagetitle = $pagename;
+// Header
+require_once('header.php');
 ?>
-
 
 <p>This keyboard layout is designed for transcribing Balochi languages based on the InPage Urdu keyboard layout. 
 It includes an on screen keyboard which can be viewed by clicking on the Keyman icon 
 and selecting the On Screen Keyboard menu item. Similar keyboards are also available for other desktop and mobile platforms.</p>
+
+<h1>Keyboard Layout</h1>
 
 <h2>Desktop Keyboard Layout</h2>
 <div id='osk' data-states='default shift right-alt right-shift-alt left-ctrl'>
@@ -20,11 +20,7 @@ and selecting the On Screen Keyboard menu item. Similar keyboards are also avail
 </div>
 
 <h2>Phone Keyboard Layout</h2>
+<p>Due to the size and number of keys, some characters are hidden in the long press.
+Press and hold on the key with a little dot on the top right to reveal them.</p>
 <div id='osk-phone' data-states='default shift numeric'>
 </div>
-
-<p>For Android and iOS, you may use long press key combinations.</p>
-
-<h2>Release History</h2>
-<p>Version 1.1, released 29 February 2020</p>
-<p>Version 1.0, released 5 August 2019</p>