+++ release/b/balochi_phonetic/source/help/balochi_phonetic.php
@@ -1,70 +1,26 @@
 <?php 
-  $pagename = 'Balochi Phonetic Keyboard Help';
-  $pagetitle = 'Balochi Phonetic Keyboard Help';
-  require_once('header.php');
+$pagename = 'Balochi Phonetic Keyboard Help';
+$pagetitle = $pagename;
+// Header
+require_once('header.php');
 ?>
-
 
 <p>This keyboard layout is designed for Balochi. It includes an on screen keyboard
 which can be viewed by clicking on the Keyman icon and selecting the On Screen Keyboard
 menu item. Similar keyboards are also available for other desktop and mobile platforms.</p>
 
-<br/>
+<h1>Keyboard Layout</h1>
 
 <h2>Desktop Keyboard Layout</h2>
 <div id='osk' data-states='default shift ctrl'>
 </div>
 
 <h2>Tablet Keyboard Layout</h2>
+<p>Due to the size and number of keys, some characters are hidden in the long press.
+Press and hold on the key with a little dot on the top right to reveal them.</p>
 <div id='osk-tablet' data-states='default shift'>
 </div>
 
 <h2>Phone Keyboard Layout</h2>
 <div id='osk-phone' data-states='default shift numeric'>
 </div>
-
-<p>For Android and iOS, you may use long press key combinations.</p>
-
-
-
-<h2>History</h2>
-
-<h3>1.3.1 (12 Feb 2020)</h3>
-<ul>
-  <li>Fixed package to include png keyboard images. Remove font from touch.</li>
-</ul>
-
-<h3>1.3 (20 Jan 2020)</h3>
-<ul>
-  <li>Put feh, hah and tatweel on the shift level since they are not used in the </li>
-  <li>Balochi alphabet. In their place put fatah, kasreh and shin. Also moved the </li>
-  <li>Arabic yeh to unused shift Y. Other shifted letters moved around to accomodate. </li>
-  <li>Other minor changes. </li>
-</ul>
-
-<h3>1.2 (9 Aug 2019)</h3>
-<ul>
-  <li>Removed deadkeys for Indic numbers. Use ctrl combination instead. </li>
-  <li>Added the zero-width joiner and better visuals for the zero-width non-joiner. </li>
-  <li>Added ', ", :, and - to desktop versions. </li>
-</ul>
-
-<h3>1.1.2 (4 Mar 2019)</h3>
-<ul>
-  <li>Remove all but the bcc language code.</li>
-</ul>
-
-<h3>1.1.1 (18 Jun 2018)</h3>
-<ul>
-  <li>No change. Package version to match keyboard.</li>
-</ul>
-
-<h3>1.1 (6 Sep 2017)</h3>
-<ul>
-  <li>Keys rearranged, damma added and Persian numbers added.</li>
-</ul>
-
-<h3>1.0 (14 Aug 2017)</h3>
-<ul>
-  <li>Initial version</li>
-</ul>