+++ release/b/balochi_urdu/HISTORY.md
@@ -1,5 +1,9 @@
 Balochi Urdu Change History
 ===============================
+1.1 (25 May 2020)
+-----------------
+Put non-Balochi letters on secondary layers.
+
 1.0.1 (4 April 2019)
 -----------------
 Just use bcc-Arab language code